+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Modal dialog
     - Add new callback action `SimpleModalBorder.OPENED` work when modal has opened
+- Extras
+    - LightDarkButton: add list support l&f `light` and `dark` to auto install themes when switch
 
 ### Changed